tools/sliver-server/pyproject.toml
[project]
name = "red-run-sliver-server"
version = "0.1.0"
description = "Sliver C2 MCP server for red-run — gRPC integration with mTLS auth"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.4.1",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
